Glue reads to references.
A ReadGroupSet now contains an explicit reference to a ReferenceSet
to which its reads are aligned. The name of the ReferenceSet is either
derived from the @SQ.AS SAM header field (if it exists) or is assumed to
be 'Default' if it does not. There must be a corresponding ReferenceSet
with this name.

Other changes:

- Refactored the datadriven tests, which had not been executing for
  variants and references;
- Removed some unnecessary sanitising in the reads code;
- Updated the data download script to edit the BAM headers so we only
  include reference headers for the chromosomes we're interested in;
- Added a some new test data so that all tests work under these new more
  stringent requirements.