Updates to Ontologies and VariantAnnotationSet.
Simplified ontologies to just have a single object representing an
ontology, backed by a single file, with the data repository providing
all the other functionality.

Partial refactor of VariantAnnotations:

The present organisation of the VariantAnnotations code was difficult to
reconcile with the DB based repo refactor. This commit gives an outline
of how it could work, by changing the relationship between VariantSet
and VariantAnnotationSet from "is a" to "has a". Unfortunately I could
not complete this work, and have had to move on to other aspects. I have
therefore disabled the tests that are failing and moved on.