Initial work on SQL DB for data repo.
Created a new API based on a SQLite DB for the data repository. This is
still WIP, and is incomplete.

Changes to download data script:

- Move the script to the project root so that it can access the ga4gh
  package.
- Changed the NCBI URL used to access sequence data as the existing one
  seems to have been discontinued.
- Added a --force/-f flag to force removal of any existing directories.
- Changed the download directory to contain a flat list of the files, as
  the hierarchy wasn't useful any more.
- Added the repo DB.
- Removed the checkpointing functionality. This would have been very
  complex to maintain now that we are using a DB rather than just
putting files in a specific location. The main reason for including it
has gone away in any case, as htslib should be much more reliable now.