Added ontology add/delete/list functionality.
The SQL schema and CLI used the term Ontology rather than
OntologyTermMap because it seems that the current approach is quite
limited and will need to be changed. This seems more forward compatible,
since we don't want to affect uses by making them change CLI syntax or
make backwards incompatible changes to the schema.

Conflicts:
	ga4gh/cli.py
	tests/unit/test_repo_manager.py