Improve response generation efficiency.
To remove a major performance bottleneck, this changes the way we
determine when the response buffer is full to use the size of the binary
encoded protobuf objects. The documentation for the MAX_RESPONSE_LENGTH
is also modified slightly.

Also fixed a logic bug in the variant annotations simulator, which had a
test to required that transcript_effects could not be empty, but the
generator have a 1/5 probability of producing this.